+++ app/Http/Controllers/TaskController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\{Board, Task};
+use Illuminate\Support\Facades\Auth;
 
 class TaskController extends Controller
 {
@@ -16,6 +17,11 @@ class TaskController extends Controller
      */
     public function store(Request $request, $bid)
     {
+        $board = Board::find($bid);
+        if ($board->user_id != Auth::user()->id) {
+            echo "access denied";
+            return; // 403
+        }
         $task = new Task;
         $task->construct($request->task, $bid);
         if ($task->save()) {
@@ -36,14 +42,19 @@ class TaskController extends Controller
     public function edit(Request $request, $bid, $tid)
     {
         //
-        $task = Task::find($tid);
-        if(empty($request->done)) {
-            $task->done = false;
+        $board = Board::find($bid);
+        if ($board->user_id != Auth::user()->id) {
+            echo "access denied";
+            return; // 403
         }
-        else if($request->done == 'on') {
+
+        $task = Task::find($tid);
+        if (empty($request->done)) {
+            $task->done = false;
+        } else if ($request->done == 'on') {
             $task->done = true;
         }
-        
+
         if ($task->save()) {
             return redirect('/boards/' . $bid);
         }
@@ -62,6 +73,11 @@ class TaskController extends Controller
     public function update(Request $request, $bid, $tid)
     {
         # code...
+        $board = Board::find($bid);
+        if ($board->user_id != Auth::user()->id) {
+            echo "access denied";
+            return; // 403
+        }
         $task = Task::find($tid);
         $task->text = $request->task;
 
@@ -83,6 +99,11 @@ class TaskController extends Controller
     public function destroy(Request $request, $bid, $tid)
     {
         //
+        $board = Board::find($bid);
+        if ($board->user_id != Auth::user()->id) {
+            echo "access denied";
+            return; // 403
+        }
         $task = Task::find($tid);
         if ($task) {
             $task->delete();
